mysql_mongo

+++ app.js
@@ -17,6 +17,7 @@ const express =require('express'),
         flash = require('connect-flash'),
         User=require('./models/user'),
         Container = require('./models/container'),
+        ip = require('ip');
         middlewareObj = require('./middleware/index');
 
 var upload=multer({dest:'uploads/'});
@@ -79,7 +80,7 @@ app.get('/login',function(req,res){
 app.post('/login', passport.authenticate('local',
 {
     
-    successRedirect:'/containers',
+    successRedirect:'/home',
     failureRedirect:'/'
 }
 ),function(req,res){
@@ -92,6 +93,9 @@ app.get('/logout',function(req,res){
     res.redirect('/');
 });
 // middlwware
+app.get('/home',middlewareObj.isLoggedIn,(req,res)=>{
+    res.render('index');
+})
 app.get('/containers', middlewareObj.isLoggedIn,async (req,res)=>{
     console.log(req.user);
     Container.find({'owner.id':req.user._id}, (err,containers)=>{
@@ -103,6 +107,110 @@ app.get('/containers/new', middlewareObj.isLoggedIn,(req,res)=>{
     console.log(req.user);
     res.render('form');
 });
+app.get('/containers/mysql',(req,res)=>{
+    res.render('mysql');
+});
+app.get('/containers/php',(req,res)=>{
+    res.render('php');
+});
+app.get('/containers/mongo',(req,res)=>{
+    res.render('mongo');
+});
+app.post('/containers/mongo',async(req,res)=>{
+    var name = req.body.name;
+    var port = Math.floor(27017+Math.random()*1000);
+    var image_name = 'mongo';
+    var resp = await shell.exec('docker run --name '+name+' --network newnet -p '+port+':27017 -d '+image_name);
+    var newContainer = {
+        container_name:name,
+        container_id:''+resp,
+        image_name:image_name,
+        port: port,
+        // image_id:image_id.stdout,
+        owner:{
+            id:req.user._id,
+            username:req.user.username
+        }
+    }
+    User.findById(req.user._id,(err,user)=>{
+        if(err){
+            console.log(err);
+        }
+        else{
+            Container.create(newContainer, (err,container)=>{
+                console.log(container);
+                user.containers.push(container._id);
+                user.save();
+                console.log(user);
+                res.redirect('/containers');
+            });
+        }
+    })
+});
+app.post('/containers/mysql', async (req,res)=>{
+    var name=req.body.name;
+    var port = Math.floor(3307 + Math.random() * 9000);
+    var image_name='mysql:5'
+    var resp = await shell.exec('docker run --name '+name+' --network newnet -p '+port+':3306 -e MYSQL_ROOT_PASSWORD=root -d '+image_name);
+    var newContainer = {
+        container_name:name,
+        container_id:''+resp,
+        image_name:image_name,
+        port: port,
+        // image_id:image_id.stdout,
+        owner:{
+            id:req.user._id,
+            username:req.user.username
+        }
+    }
+    User.findById(req.user._id,(err,user)=>{
+        if(err){
+            console.log(err);
+        }
+        else{
+            Container.create(newContainer, (err,container)=>{
+                console.log(container);
+                user.containers.push(container._id);
+                user.save();
+                console.log(user);
+                res.redirect('/containers');
+            });
+        }
+    })
+    
+})
+app.post('/containers/php', async (req,res)=>{
+    var name=req.body.name;
+    var mysql_container = req.body.mysql_container;
+    var port = Math.floor(3000 + Math.random() * 9000);
+    var image_name='phpmyadmin'
+    var resp = await shell.exec('docker run --name '+name+' --network newnet --link '+mysql_container+':db -p '+port+':80  -d '+image_name);
+    var newContainer = {
+        container_name:name,
+        container_id:''+resp,
+        image_name:image_name,
+        port: port,
+        // image_id:image_id.stdout,
+        owner:{
+            id:req.user._id,
+            username:req.user.username
+        }
+    }
+    User.findById(req.user._id,(err,user)=>{
+        if(err){
+            console.log(err);
+        }
+        else{
+            Container.create(newContainer, (err,container)=>{
+                console.log(container);
+                user.containers.push(container._id);
+                user.save();
+                console.log(user);
+                res.redirect('/containers');
+            });
+        }
+    })  
+})
 app.post('/containers/new', middlewareObj.isLoggedIn,async (req,res)=>{
     var url = req.body.git_url;
     var name=req.body.name;
@@ -126,8 +234,9 @@ app.post('/containers/new', middlewareObj.isLoggedIn,async (req,res)=>{
     await shell.cp('-r',`/home/${USER}/Documents/RemoteServer/Dockerfile`,newpath);
     await shell.cd(newpath);
     
-    var image_id = await shell.exec('docker build --build-arg '+env+' -t '+image_name+' .');
-    var resp = await shell.exec('docker run --name '+name+' -p '+port+':8080 -d '+image_name);
+    // var image_id = await shell.exec('docker build --build-arg '+env+' -t '+image_name+' .');
+    var image_id = await shell.exec('docker build -t '+image_name+' .');
+    var resp = await shell.exec('docker run --name '+name+' --network newnet -p '+port+':8080 -d '+image_name);
     console.log('container = ' +resp);
     var newContainer = {
         github_link:url,
@@ -163,6 +272,14 @@ app.get('/containers/:id', middlewareObj.isLoggedIn, async (req,res)=>{
         res.render('container',{container:container});
     });
 });
+//getting logs from the containers
+app.get('/containers/:id/logs',async(req,res)=>{
+    Container.findById(req.params.id,async(err,container)=>{
+        var logs = await shell.exec('docker logs '+container.container_name);
+        console.log( 'the ip address is '+ip.address());
+        res.send(logs);
+    });
+});
 app.put('/containers/:id', middlewareObj.isLoggedIn, async (req,res)=>{
     // console.log(req.param.id);
     Container.findById(req.params.id, async (err,container)=>{
@@ -190,16 +307,18 @@ app.put('/containers/:id', middlewareObj.isLoggedIn, async (req,res)=>{
 app.delete('/containers/:id',middlewareObj.isLoggedIn, async (req,res)=>{
     // console.log(req.param.id);
     Container.findByIdAndDelete(req.params.id, async (err,container)=>{
-        console.log(container);
-        var url = container.github_link;
+        console.log(container + "for deletion *********");
         const path ='/home/akash/uploads/';
         await shell.exec('docker stop '+container.container_name);
         await shell.exec('docker rm '+container.container_name);
-        await shell.exec('docker rmi '+container.image_name);
-        a = url.split('/');
-        b = a[a.length - 1].split('.')[0]
-        var newpath=path+b;
-        await shell.exec('rm -rf '+newpath);
+        if (container.github_link){
+            await shell.exec('docker rmi '+container.image_name);
+            var url = container.github_link;
+            a = url.split('/');
+            b = a[a.length - 1].split('.')[0]
+            var newpath=path+b;
+            await shell.exec('rm -rf '+newpath);
+        }
         
         User.findById(req.user._id,(err,user)=>{
             var index = user.containers.indexOf(container._id);
